requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
black==21.11b1
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
Deprecated==1.2.13
distro==1.6.0
Django==3.2.9
djangorestframework==3.12.4
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flake8==3.9.2
flake8-django==1.1.2
idna==3.3
jsonschema==3.2.0
kombu==5.2.2
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
paramiko==2.8.0
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.23
psycopg2==2.9.2
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyNaCl==1.4.0
pyright==0.0.13
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==4.0.2
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.13.3
